Update Serverless to fix security vulnerability in Axios
See https://github.com/serverless/serverless/issues/8712 for
details of the security vulnerability.

This update to Serverless has necessitated us installing "ajv"
as a dev dependency. "ajv" is a peer dependency of one of Serverless'
dependencies, and shouldn't normally require us to install it
(because Serverless itself does), however there is a long-standing
bug in npm that leads to npm being unable to resolve peer
dependencies in certain circumstances, this being one of them.
More details at
https://github.com/serverless/serverless/issues/8727
and
https://github.com/npm/npm/issues/19877