TypeScript support excluding tests
Excluding tests for now since it will involve adding Babel due to
Jest compatibility issues.

A refactor / removal of the error handling in handler which is more
simpler due to it being quite difficult to have safe types in.

Two distinct course types, one from Funnelback and our API. Previously
this was somewhat hidden as it all appeared as one.